Add coverage for flaky branch
A small percentage of test runs, the Atomics optimization on the
main thread side isn’t covered by the existing tests because the
timing doesn’t work out correctly, rendering our CI flaky.

The added test aims to correct that by coordinating the timing of tasks
in a way that ensures that the optimization is covered.