Wait for Workers to be ready before considering them as task candidates
Refactor the Worker pool management in order to better accommodate the
async nature of how they are started, and use that to wait until the
target module has been loaded before considering the Worker as a
task processing candidate in order to improve performance.

This also includes bugfixes for working around a crash in Node.js core.