Add Piscina.prototype.run(), deprecate runTask()
The `runTask()` method uses positioned arguments for various
inputs, making the signature and the argument validation
fairly complicated. This makes it more difficult to add new
options later.

This commit introduces a new `run()` function that accepts an
options object instead, e.g.

```js
piscina.run({ a: 1 }, { filename: '...', signal: ... });
```

This will be more scalable in the long run if new options need
to be added.

This marks the `runTask()` variant as deprecated.