feat: use native Node.js histogram support
The histogram related dependencies (`hdr-histogram-js`/`hdr-histogram-percentiles-obj`)
have been replaced with the native support present in Node.js. This allows the package
to have no required dependencies (with one optional for nice support on linux).
The public API surface was kept in place to prevent breaking changes for users. This
required a small amount of additional code to mainly transform the percentiles to the
previous structure.