feat: add option to disable run/wait time recording
An additional `Piscina` constructor option is now available named `recordTiming`
which can be used to control the underlying recording of wait and run times
via Node.js histograms. By default this option is `true` to retain existing behavior
but can be set to `false` to disable recording. The `Piscina` instance properties
`runTime`, `waitTime`, and `utilization` will not be available if this option is
disabled.